refactor: split SQLite SQL generation modules
