+++ src/infra/adapters/sqlite/sql/mod.rs
@@ -1,0 +1,11 @@
+mod ddl;
+mod dialect;
+mod literal;
+mod metadata;
+
+pub(super) use literal::{PREVIEW_TRANSPORT_UNISTR_PREFIX, sqlite_nul_text_sentinel};
+pub(super) use metadata::{
+    TableMetadataQueryMode, build_preview_query, has_virtual_tables_query,
+    is_table_list_unavailable, legacy_user_tables_query, preview_metadata_query,
+    table_list_required_error, table_metadata_query, table_signatures_query, user_tables_query,
+};
